feat(design-system):update icon style
