+++ libs/ui/design-system/src/lib/components/IconPicker/IconPicker.tsx
@@ -10,7 +10,7 @@ import { Popover } from "../Popover/Popover";
 import { LABEL_CLASS } from "../constants";
 import "./IconPicker.scss";
 import { EnumIconFamily, Icon } from "../Icon/Icon";
-import { EnumTextColor } from "../Text/Text";
+import { ICONS } from "./constants";
 
 const CLASS_NAME = "amp-icon-picker";
 
@@ -20,157 +20,6 @@ export type Props = {
   label?: string;
   iconOnlyMode?: boolean;
 };
-
-export const ICONS = [
-  "beaker",
-  "cloud-download",
-  "code",
-  "database",
-  "file-binary",
-  "file-code",
-  "graph",
-  "key",
-  "keyboard",
-  "law",
-  "light-bulb",
-  "link",
-  "mail",
-  "package",
-  "pulse",
-  "rocket",
-  "ruby1",
-  "server",
-  "settings",
-  "shield",
-  "tag",
-  "telescope",
-  "tools",
-  "1password",
-  "algolia",
-  "amazonaws",
-  "amd",
-  "android",
-  "angular",
-  "ansible",
-  "apache",
-  "apacheairflow",
-  "apachecordova",
-  "apacheflink",
-  "apachekafka",
-  "apachenetbeanside",
-  "apacheopenoffice",
-  "apacherocketmq",
-  "apachespark",
-  "apple",
-  "auth0",
-  "azureartifacts",
-  "azuredevops",
-  "azurepipelines",
-  "babel",
-  "bitbucket",
-  "bitcoin",
-  "cassandra",
-  "cisco",
-  "cloudflare",
-  "codepen",
-  "codesandbox",
-  "coffeescript",
-  "csharp",
-  "css3",
-  "deno",
-  "django",
-  "docker",
-  "dot-net",
-  "drupal",
-  "elastic",
-  "flask",
-  "flutter",
-  "git",
-  "github",
-  "gitlab",
-  "gitpod",
-  "gitter",
-  "go",
-  "google",
-  "gradle",
-  "grafana",
-  "graphcool",
-  "graphql",
-  "gulp",
-  "haskell",
-  "helm",
-  "heroku",
-  "html5",
-  "ionic",
-  "ios",
-  "java",
-  "javascript",
-  "jekyll",
-  "jenkins",
-  "jest",
-  "jira",
-  "jquery",
-  "jsfiddle",
-  "kotlin",
-  "kubernetes",
-  "laravel",
-  "linux",
-  "linuxfoundation",
-  "mailchimp",
-  "microsoft",
-  "microsoftazure",
-  "mongodb",
-  "mysql",
-  "neo4j",
-  "netlify",
-  "next-dot-js",
-  "nginx",
-  "node-dot-js",
-  "npm",
-  "nuget",
-  "nuxt-dot-js",
-  "opensuse",
-  "openvpn",
-  "oracle",
-  "php",
-  "postgresql",
-  "postman",
-  "powershell",
-  "prettier",
-  "python",
-  "pytorch",
-  "pyup",
-  "r-language",
-  "rabbitmq",
-  "rails",
-  "raspberrypi",
-  "react",
-  "reactos",
-  "redhat",
-  "redis",
-  "redux",
-  "rss",
-  "ruby",
-  "rust",
-  "sass",
-  "scala",
-  "snyk",
-  "sonarqube",
-  "stripe",
-  "svg",
-  "swagger",
-  "swift",
-  "tailwindcss",
-  "terraform",
-  "twilio",
-  "x-social",
-  "typescript",
-  "unity",
-  "visualstudiocode",
-  "webpack",
-  "whatsapp",
-  "youtube",
-];
 
 export const IconPicker: React.FC<Props> = ({
   selectedIcon,
@@ -232,7 +81,7 @@ export const IconPicker: React.FC<Props> = ({
               buttonStyle={EnumButtonStyle.Text}
               onClick={() => setIsOpen(!isOpen)}
               type="button"
-              icon="icon"
+              icon="plugin"
             />
           ) : (
             <Button
