+++ libs/ui/design-system/src/lib/components/VerticalNavigation/VerticalNavigationItem.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback, useEffect } from "react";
 import { NavLink, useRouteMatch } from "react-router-dom";
-import { Icon } from "../Icon/Icon";
+import { EnumIconFamily, Icon } from "../Icon/Icon";
 import classNames from "classnames";
 import { VerticalNavigation } from "./VerticalNavigation";
 import "./VerticalNavigationItem.scss";
@@ -10,6 +10,7 @@ import { Button, EnumButtonStyle } from "../Button/Button";
 export type Props = {
   children: React.ReactNode;
   icon?: string;
+  iconFamily?: EnumIconFamily;
   to: string;
   className?: string;
   childItems?: React.ReactNode;
@@ -26,6 +27,7 @@ export function VerticalNavigationItem({
   className,
   childItems,
   expandable,
+  iconFamily,
   onExpand,
 }: Props) {
   const [expanded, setExpanded] = React.useState(false);
@@ -62,7 +64,7 @@ export function VerticalNavigationItem({
       >
         <FlexItem
           itemsAlign={EnumItemsAlign.Center}
-          start={icon && <Icon icon={icon} size="small" />}
+          start={icon && <Icon icon={icon} size="small" family={iconFamily} />}
           end={
             expandable && (
               <Button
